# Optimize campaigns index page performance - 15-20x faster